Dont create log dir when threshold config=0
When threshold config = 0 the application was still trying to create the log sub directory. Leading to error if we dont create the logs parent directory.

I would also advice to use 0 as the default config for threshold. It will allow slightly easyest demployement because you dont have to create+chmod the log directory.